Naive implementation spiked out
Fully functioning, naive implementation of the hash approach using the built-in python dictionary! The test RNA sequences defined include 1 hit for each DCE, except 712-A-r which should have 2. The current results reflect this accurately.
There are remnants of the bloom filter approach I started, but this first pass was too simple to get functioning properly for this and isn't yet worth putting more time into.