Add argument parser
Argument parser added to allow user to pass in the filenames for the target and query (i.e. DNA and RNA) sequence files they have. If no arguments are provided, it will default to the current test files (DCEs.fasta and RNAseqs.fasta). Also displays summary of program use when --help argument is input.