Reduce RNA overhead, complete optimization
Finished removing and debugging the interim objects for the RNA sequences that were drastically increasing overhead. Also had to fix the bug with concatenating the RNA across lines where I forgot to wipe the string between sequences and the whole file was getting concatenated together... whoops.
Now runs in <3 minutes on the actual files that took >30 with bowtie!