Increase file reading and revcomp performance
Switched RNA file to be read in in chunks instead of trying to get the whole giant thing, which yielded a small performance increase.
Also changed the reverse complement search to use the str.translate method, which uses a C level lookup. This has improved runtime by a bit more than 8000% - it still takes a little over an hour to run on the full ~9GB SRR1099957 file, but that is a serious improvement over the 8+ hours it took previously (on my local machine).