Confirm results, add tools for further validation
Debug completed - as I suspected, it was an issue with hash access weirdness. After fixing it, the test DCE (4-S-l) has the same number of hits from bowtie, count, and grep.
Instead of testing on the full RNAseq file (SRR1099957), I used part 4 (as output from the fasta splitter with part size as specified in earlier comments) and removed the extra linebreaks in the sequences using remove_lines.py. Then running search4.py gives you the number of hits of the test DCE 4-S-l, split into forward and reverse (160 hits - 55 of them forward, 105 reverse comp). Grep reports the same number when searching for both the forward and reverse complements of the 4-S-l sequence, i.e.:
grep -o 'GAAGCCATCAGGACGTTCACGGCCGAGGTGGT' SRR1099957.4.test.fasta | wc -l
grep -o 'ACCACCTCGGCCGTGAACGTCCTGATGGCTTC' SRR1099957.4.test.fasta | wc -l
After fixing the problems with hash access in count.py, it now reports 160 hits as well! I did a spot check with 30-A-l and 267-S-l as well, and count and grep match on the number of hits for those as well. (The output of all hits found in the partial file is uploaded in test.out, so George can use this as-is for rust validation)

Finally, parseSAM.pl has a slight modification in the regex to filter the bowtie results to only exact matches, now that I found out that it attempts to find alignments with 0 AND 1 mismatches before trying the multiseed heuristic. Final and complete validation of results will come shortly - what remains is filtering bowtie's full SRR1099957 file results with this new parser, and running count on the full file to compare the output, just to make sure the results of part 4 generalize to the full sequence file.