Provide working test generator and small refactor of count
Count.py refactored to actually move the separate chunks handling the DNA, RNA, etc to their own functions instead of being in main. Also added one more function, make_results_hash, that reports the number of hits for ALL items in the DCE hash, not just those that have 1+ hits. This enables it to be used for comparison in test_generator.

Also obviously adding the test_generator - right now as far as I know it's done and working, but obviously can be reworked more... right now it just prints "Pass!" on the command line if it passes. This file has several functions to generate the random DNA sequences and random RNA sequences, as well as selecting which of those sequences will be inserted into the RNA seqs. (N is total number of DNA sequences to be generated, while K is the number that should be inserted in RNA.) After all the generation/insertion is done, the check_grep function outputs the number of hits yielded from running grep.
At the end it finally calls count functions on the generated test DNA and RNA, then compares the result hash to the prior result hash from grep. Hopefully it'll print "Pass!" for you too... and here is where we have room to start incorporating an actual test suite.