Report total hit counts for each result
Super naive implementation of feature to report total hit counts added. For each search result reported, it now also prints out the running total of hits for each DCE. This is accomplished by adding a "hits" field to the CompressedSeq struct, making the needle mutable, and incrementing its hits number every time there is a match while searching.
Adaptations are still required to create results matching the python version, i.e. writing out to a results file that contains the name of each DCE that had >= 1 hit, and the total number of hits it had.