Begin hash implementation (for DCEs), add bio library
Hash implementation started, using the Rust-Bio library for reading in the DCE fasta file (temporarily, until our implementation has been debugged). This required refactoring of the fasta.rs file, which has been renamed to seqloader to resolve shared name conflicts with Rust-Bio.
The DCE hash was implemented using the multimap library, which is a small wrapper around the hashmap that allows us to put multiple values (IDs) in for each sequence without additional handling. As such, support of non-unique sequences has already been added. These changes  (fasta reading/moving away from the Seq struct) still need to be implemented for the RNAseq file, as well as changes to Search to support checking against the DCE hash.
And yes... main is (temporarily) a bit of a mess.