Implement search with DCE hash
Perform the search while continuing to use the Seq struct for the RNA sequences, but using a hash (instead of vec) for the DCEs.
DCE compression is now happening at the same step as loading them from the file - they are immediately bitpacked and only stored in their bitpacked forms. The bitpacked hash is then passed into the Search function (but still currently has to be cloned, which isn't ideal). This enables just checking to see if the current RNA 32mer exists in the hash instead of using a needle index to walk through the vector.

It's "working" right now, except when it finds the first hit it gets stuck in an infinite loop instead of stepping forward to the next RNA 32mer...