Finish revcomp debugging!
Fixed the last bug and a manual check confirms that matches for 4-S-l are now being reported correctly - 55 for forward, 145 for revcomp, for the correct 160 hits total. Mapping the integers (representing the bitpacked sequence) back to the DCE_ID will need to be done before further verification can be done.