Add hits vector to hash, and method to increment
To get total hits working, both the field (the u16 vector as discussed) and a method to increment the hits have been added to Hash.
In Search, when a hit has been found, after adding the search result to the vector, the total hit count for that DCE has also been incremented by 1.

The work that remains is accessing that part of the Hash from main, where the total hit counts was previously reported.