Both paths for adding hits array
Methods added in hash to increment hits, return the hash value, and print the hit(s).
Creation of needles hash moved to main, in an attempt to access the hits array for printing results at the end, but this has created ownership issues because we can't implement the copy trait on the Hash struct. Changed to pass-by-reference (and added a print function into Search) to try to get around this, but can't get results to print out properly...
For the other approach: In main it also currently creates a separate hits vector. Search Result has been modified to also return the hit index (aka DCE index) that allows us to increment that separate hits vector, although the results are currently reported tied to the hash value of the DCE instead of the sequence or ID.