Implement total hit counts and ID mapping
Both features are implemented and working as of now. Debugging done to get the total hit counts working properly (output has been verified against the python/grep output and confirmed to be correct).
For mapping back to the DCE_ID (instead of reporting by bitpacked sequence), a new file/structure was added - a hashmap very very similar to the one made to hold the bitpacked sequence for searching, but this one holds another vector of Strings (for the ID) instead of a hits vector. Several unnecessary functions have also been removed. This is filled in at the same time the DCEs are initially being read in, then just used at the very end to map everything back together and report hit counts for each DCE_ID.