Add non-unique sequence support
Initial non-unique sequence support added, with results being reported for duplicate sequences (i.e. reporting number of hits for both 664-S-l and 665-S-l, instead of just the first one).
Mostly working, but some slight debugging to do; specifically the hits for 1128-A-l and 1272-A-r are no longer being reported, so I need to do some investigating into why that is. The known duplicates are being correctly reported with matching number of hits, however.