Debugging of DCE_ID misreporting
Pushing up small changes to enable more debugging of the DCE_ID misreporting. (Such as having output.txt include the index of the dce, confirm the hashmap is being built correctly at beginning, etc.) Trying to get the index of each DCE from the hashmap is still not working due to conflicting immutable and mutable borrows at line 97 vs 100 (currently).

The remaining bug is two misidentification problems. It is currently reporting the results for 1128-A-l under the name 1253-A-l, and 1272-A-l is being reported under the name "0". All other results are confirmed correct, and the number of hits is verified.