Update post-processing methods, verify results
Switched from the custom Hashmap structure to using Multimap, since prior work made it possible to just use the compressed sequence to retrieve the number of hits for that sequence. Reported sequences and number of hits have been compared to the verified Python output and the weird bug is finally routed! Results are 100% correct.

In addition, updated the output file to print out the proper information - i.e. DCE ID, then number of hits for that DCE on each line. The first line in the file also now reports which RNA file it was run on, which will be necessary soon for doing the large-scale analysis.

Final runtime clocks in at ~130 seconds per run.