Merge debugged hits-id into master branch (#21)
* Fasta to fastq reader changes

Comments and beginning changes made for converting the fasta reader to a fastq reader, now that the SRA toolkit doesn't appear to allow us to dump the SRA data into fasta format

* Debugging of DCE_ID misreporting

Pushing up small changes to enable more debugging of the DCE_ID misreporting. (Such as having output.txt include the index of the dce, confirm the hashmap is being built correctly at beginning, etc.) Trying to get the index of each DCE from the hashmap is still not working due to conflicting immutable and mutable borrows at line 97 vs 100 (currently).

The remaining bug is two misidentification problems. It is currently reporting the results for 1128-A-l under the name 1253-A-l, and 1272-A-l is being reported under the name "0". All other results are confirmed correct, and the number of hits is verified.

* Add test to demonstrate potential bug

* Update post-processing methods, verify results

Switched from the custom Hashmap structure to using Multimap, since prior work made it possible to just use the compressed sequence to retrieve the number of hits for that sequence. Reported sequences and number of hits have been compared to the verified Python output and the weird bug is finally routed! Results are 100% correct.

In addition, updated the output file to print out the proper information - i.e. DCE ID, then number of hits for that DCE on each line. The first line in the file also now reports which RNA file it was run on, which will be necessary soon for doing the large-scale analysis.

Final runtime clocks in at ~130 seconds per run for a 1GB RNAseq file. 

* Clean up codebase

Removed the now-unneeded Hashmap file, as well as all of the commented out bits of code that were used for debugging.

Co-authored-by: George Lesica <george@lesica.com>