Change to SAM format
Output has been changed to match the SAM format, although much of the data has not been filled in. Instead of outputting just the given DCE and how many hits it got, the DCE has to be printed once for each hit.
Currently most of the columns are filled in with values that denote that value is not to be used, although in the future it may be worth changing the flag to note which hits were reverse complement